clean up dgcn code
