+++ torch_points3d/modules/DGCN/helpers.py
@@ -2,16 +2,15 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.autograd import Variable
-# from pdb import set_trace as brk
 
 def get_edge_features_xyz(x, pc, k, num=-1):
     """
     Args:
-        x: point cloud [B, dims, N]
+        x: point features [B, dims, N]
+        pc: point cloud [B, dims, M]
         k: kNN neighbours
     Return:
-        [B, 2dims, N, k]
-        idx
+        ([B, 2dims, N, k], [B, 2dims, M, k])
     """
     B, dims, N = x.shape
 
@@ -58,7 +57,7 @@ def get_edge_features_xyz(x, pc, k, num=-1):
 def get_edge_features(x, k, num=-1):
     """
     Args:
-        x: point cloud [B, dims, N]
+        x: point features [B, dims, N]
         k: kNN neighbours
     Return:
         [B, 2dims, N, k]    
@@ -94,37 +93,3 @@ def get_edge_features(x, k, num=-1):
     ee = torch.cat([central, neighbors-central], dim=1)
     assert ee.shape == (B, 2*dims, N, k)
     return ee
-
-class ChamferLoss(nn.Module):
-
-	def __init__(self):
-		super(ChamferLoss, self).__init__()
-		self.use_cuda = torch.cuda.is_available()        
-
-	def forward(self,preds,gts):
-		P = self.batch_pairwise_dist(gts, preds)
-		mins, _ = torch.min(P, 1)
-		loss_1 = torch.sum(mins)
-		mins, _ = torch.min(P, 2)
-		loss_2 = torch.sum(mins)
-
-		return loss_1 + loss_2
-
-
-	def batch_pairwise_dist(self,x,y):
-		bs, num_points_x, points_dim = x.size()
-		_, num_points_y, _ = y.size()
-		xx = torch.bmm(x, x.transpose(2,1))
-		yy = torch.bmm(y, y.transpose(2,1))
-		zz = torch.bmm(x, y.transpose(2,1))
-		if self.use_cuda:
-			dtype = torch.cuda.LongTensor
-		else:
-			dtype = torch.LongTensor
-		diag_ind_x = torch.arange(0, num_points_x).type(dtype)
-		diag_ind_y = torch.arange(0, num_points_y).type(dtype)
-		#brk()
-		rx = xx[:, diag_ind_x, diag_ind_x].unsqueeze(1).expand_as(zz.transpose(2,1))
-		ry = yy[:, diag_ind_y, diag_ind_y].unsqueeze(1).expand_as(zz)
-		P = (rx.transpose(2,1) + ry - 2*zz)
-		return P
